Moverd compare functions to helpers
